requirements.txt
python-dotenv
pyTelegramBotAPI
schedule
